+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: hokutosuzuki <marvin@42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/16 18:03:52 by hokutosuz         #+#    #+#             */
-/*   Updated: 2021/11/02 17:41:20 by hokutosuz        ###   ########.fr       */
+/*   Updated: 2021/11/04 10:25:38 by hokutosuz        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ char	*ft_strrchr(const char *s, int c)
 {
 	char	*p;
 	char	*p2;
-	int		i;
+	size_t	i;
 
 	p = (char *)s;
 	p2 = NULL;
